Squashed 'libraries/AtSamd/' changes from 543353e..aa8a87e
aa8a87e Return the number of MS for the period.
REVERT: 543353e Return the number of MS for the period.

git-subtree-dir: libraries/AtSamd
git-subtree-split: aa8a87e2bf3c8ab767a27beaf992050f8b805804

+++ wdt.cpp
@@ -83,7 +83,7 @@ uint16_t wdt_get_period_length_in_ms(uint8_t period) {
     }
 }
 
-uint8_t wdt_enable(uint8_t period) {
+uint16_t wdt_enable(uint8_t period) {
     wdt_initialize();
 
     // First disable the watchdog so its registers can be changed.
